requirements.txt
numpy==1.24.3
pandas==2.0.3
biopython==1.81
tqdm==4.65.0
snakemake==6.0.0